% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_keybase.R
\name{send_keybase}
\alias{send_keybase}
\title{Send a message via Keybase}
\usage{
send_keybase(user, message = NULL, file = NULL)
}
\arguments{
\item{user}{The recipient's Keybase username.}

\item{message}{A character string message to send.}

\item{file}{Path to an image (optional). If supplied, the file is uploaded.}
}
\value{
Invisibly returns TRUE if successful.
}
\description{
Send a message via Keybase
}
